# Improve accepted JSON values in JsonObjectParser