--- a/akka-stream-tests/src/test/scala/akka/stream/scaladsl/JsonFramingSpec.scala
+++ b/akka-stream-tests/src/test/scala/akka/stream/scaladsl/JsonFramingSpec.scala
@@ -276,8 +276,6 @@ class JsonFramingSpec extends AkkaSpec {
       result.futureValue shouldBe Seq("""[ [ 1, 2], [ 3, 4 ] ]""", """[ 5, [ 6, 7 ], 8 ]""")
     }
 
-
-
     "emit single json element from string" in {
       val input =
         """| { "name": "john" }