… akka-bench-jmh/src/main/scala/akka/stream/JsonFramingBenchmark.scala
… akka-stream/src/main/scala/akka/stream/scaladsl/Framing.scala