This PR fixes #25673 by making the JsonObjectParser framing engine able to also accept:

- arrays of scalar Json values (strings, booleans, numbers)
- linebreak or comma-separated scalar Json values 

The tests are expanded to reflect this improvement.

In terms of implementation, the increased complexity led me to extract the state machine logic (formerly in ```def proceed```) into independent private state objects (which are stored within the companion in order to avoid creating tons of instances at each new stream). The JsonObjectParser will now behave very differently according to the following keys:
  - if the first non-whitespace character is a ```[```, then the parser assumes the "JSON Array" encoding and will enforce a single value per item, separated by commas, and optionally terminated with a single ```]``` (it will not complain if that is missing).
  - otherwise, it runs in "separator-separated mode". This mode is subdivided in comma or linebreak-separated mode, depending in which separator is found after the first value. 

Furthermore, it was discovered as part of running the JMH Benchmark that the previous implementation of JsonObjectParser was able to accept streams of unseparated JSON *Object* values, i.e. it did correctly frame the following stream:
```json
  {"foo":1}{"foo":2}{"foo":3}
```
even though there is neither a comma or a line break between objects.
The tests are further expanded to reflect this. All three of "yet undetermined separator-separated", "comma-separated" and "linebreak-separated" modes accept the above sequence.

In terms of performance, at at **-wi 5 -i 5** iterations:

**Before**, (https://github.com/akka/akka/commit/16438fcb860f932aec4187c27053884dcf2d9fe5)
```text
Benchmark                                     Mode  Cnt        Score        Error  Units
JsonFramingBenchmark.counting_1              thrpt    5  2084711,811 ± 207865,821  ops/s
JsonFramingBenchmark.counting_long_document  thrpt    5      186,252 ±    116,947  ops/s
JsonFramingBenchmark.counting_offer_5        thrpt    5  2306649,473 ± 471682,266  ops/s
```
**After**:
```text
Benchmark                                     Mode  Cnt        Score        Error  Units
JsonFramingBenchmark.counting_1              thrpt    5  3418384,183 ± 481082,430  ops/s
JsonFramingBenchmark.counting_long_document  thrpt    5      728,951 ±     30,755  ops/s
JsonFramingBenchmark.counting_offer_5        thrpt    5  2772735,430 ± 321462,765  ops/s
```

When running at a high amount of iterations, it seems something strange kicks in (running with **-wi 30 -i 30**):
**Before**
```text
JsonFramingBenchmark.counting_1              thrpt   30  2565875,250 ± 24452,637  ops/s
JsonFramingBenchmark.counting_long_document  thrpt   30      190,048 ±     2,777  ops/s
JsonFramingBenchmark.counting_offer_5        thrpt   30  2188289,936 ± 32093,784  ops/s
```

**After**
```text
Benchmark                                     Mode  Cnt        Score       Error  Units
JsonFramingBenchmark.counting_1              thrpt   30  2409937,623 ± 10269,970  ops/s
JsonFramingBenchmark.counting_long_document  thrpt   30      700,063 ±    15,331  ops/s
JsonFramingBenchmark.counting_offer_5        thrpt   30  2078812,214 ± 45579,564  ops/s
```
There is a noticeable drop in throughput for ```counting_1``` and ```counting_offer_5``` after the 20th warm-up iteration. It is unclear to me whether JsonObjectParser, thermal CPU throttling or the test fixture itself is the culprit (the nominal CPU frequency appeared to stay put at 3509 MHz throughout the test runs)
```text
# JMH version: 1.21
# VM version: JDK 1.8.0_171, OpenJDK 64-Bit Server VM, 25.171-b11
# scala 2.12.6
```