@cchepelov 2018-10-12
Thanks for the review @johanandren and apologies for the weeks in a tunnel.

Regarding the GC, will benchmark. I tried hard to avoid creating objects at runtime (there are a number of objects that get created as part of the stream's init, and a few that can get created as part of the state stack, but there's unconventional recycling for this). 

Will report, I understand this is an important metric to reach a decision.