@johanandren 2018-10-12
No worries, we'll be fairly busy in the team the coming weeks as well, so don't feel any pressure to hurry up from our side.