@cchepelov 2018-10-12
Post-review, revisiting performance:

master (d9a41f5e237a841b7f15e198fb5c22673b6ed1cf):
```text
[info] Benchmark                                     Mode  Cnt        Score        Error  Units
[info] JsonFramingBenchmark.counting_1              thrpt    5  2876586,577 ± 104206,810  ops/s
[info] JsonFramingBenchmark.counting_long_document  thrpt    5      216,529 ±      4,662  ops/s
[info] JsonFramingBenchmark.counting_offer_5        thrpt    5  2436713,062 ±  85255,671  ops/s
```

Initial pull request (641961ddcf8a7bfd46cee6c06722573c76b485b4)
```text
[info] Benchmark                                     Mode  Cnt        Score        Error  Units
[info] JsonFramingBenchmark.counting_1              thrpt    5  3191640,194 ± 787917,209  ops/s
[info] JsonFramingBenchmark.counting_long_document  thrpt    5      734,874 ±    128,198  ops/s
[info] JsonFramingBenchmark.counting_offer_5        thrpt    5  2754103,799 ± 384199,265  ops/s
```

Updated branch (bb456bd7d2c9b9524a35dbfe7c463e7e8967a7d6), today:
```text
[info] Benchmark                                     Mode  Cnt        Score        Error  Units
[info] JsonFramingBenchmark.counting_1              thrpt    5  3438107,284 ± 348074,002  ops/s
[info] JsonFramingBenchmark.counting_long_document  thrpt    5      750,122 ±     17,825  ops/s
[info] JsonFramingBenchmark.counting_offer_5        thrpt    5  2794009,199 ± 285732,751  ops/s
```

There seems to be a minute throughput improvement, possibly due to the improvement you suggested in patmat. I'll investigate whether it's worth expanding ` case _ if isWhitespace(input) ⇒` into `case Space|Linebreak|Linebreak2|Tab ⇒` despite the obvious maintenance drawback. 

Regarding the allocation question:
master (d9a41f5e237a841b7f15e198fb5c22673b6ed1cf):
```text
[info] JsonFramingBenchmark.counting_1                                     thrpt    5  2624091,126 ± 1219347,872   ops/s
[info] JsonFramingBenchmark.counting_1:·gc.alloc.rate                      thrpt    5      133,462 ±      62,014  MB/sec
[info] JsonFramingBenchmark.counting_1:·gc.alloc.rate.norm                 thrpt    5       56,000 ±       0,001    B/op
[info] JsonFramingBenchmark.counting_long_document                         thrpt    5      210,822 ±      13,792   ops/s
[info] JsonFramingBenchmark.counting_long_document:·gc.alloc.rate          thrpt    5        0,011 ±       0,001  MB/sec
[info] JsonFramingBenchmark.counting_long_document:·gc.alloc.rate.norm     thrpt    5       56,208 ±       0,015    B/op
[info] JsonFramingBenchmark.counting_offer_5                               thrpt    5  2477899,025 ±  115321,706   ops/s
[info] JsonFramingBenchmark.counting_offer_5:·gc.alloc.rate                thrpt    5      608,525 ±      28,318  MB/sec
[info] JsonFramingBenchmark.counting_offer_5:·gc.alloc.rate.norm           thrpt    5      270,400 ±       0,001    B/op
```

Updated branch (bb456bd7d2c9b9524a35dbfe7c463e7e8967a7d6), today:
```text
[info] JsonFramingBenchmark.counting_1                                         thrpt    5  3636627,358 ± 409547,095   ops/s
[info] JsonFramingBenchmark.counting_1:·gc.alloc.rate                          thrpt    5      132,115 ±     14,873  MB/sec
[info] JsonFramingBenchmark.counting_1:·gc.alloc.rate.norm                     thrpt    5       40,000 ±      0,001    B/op
[info] JsonFramingBenchmark.counting_long_document                             thrpt    5      764,573 ±     13,482   ops/s
[info] JsonFramingBenchmark.counting_long_document:·gc.alloc.rate              thrpt    5        0,028 ±      0,004  MB/sec
[info] JsonFramingBenchmark.counting_long_document:·gc.alloc.rate.norm         thrpt    5       40,700 ±      5,524    B/op
[info] JsonFramingBenchmark.counting_offer_5                                   thrpt    5  2942200,995 ± 102472,910   ops/s
[info] JsonFramingBenchmark.counting_offer_5:·gc.alloc.rate                    thrpt    5      209,497 ±      7,291  MB/sec
[info] JsonFramingBenchmark.counting_offer_5:·gc.alloc.rate.norm               thrpt    5       78,400 ±      0,001    B/op
```