@johanandren 2018-10-12
Alloc looks good, thanks for double checking that.

Make sure to warmup until the throughput stabilizes and then probably run a few more than 5 iterations for the actual benchmark, just to make sure we're not seeing pre-jit performance etc in there.