@cchepelov 2018-10-12
> There seems to be a minute throughput improvement, possibly due to the improvement you suggested in patmat. I'll investigate whether it's worth expanding `case _ if isWhitespace(input) ⇒` into `case Space|Linebreak|Linebreak2|Tab ⇒` despite the obvious maintenance drawback.

Not getting really obviously better numbers: 
```text
[info] Benchmark                                     Mode  Cnt        Score        Error  Units
[info] JsonFramingBenchmark.counting_1              thrpt    5  3438107,284 ± 348074,002  ops/s     → 3478521,210 ± 266340,707  ops/s
[info] JsonFramingBenchmark.counting_long_document  thrpt    5      750,122 ±     17,825  ops/s     → 758,564 ±     37,204  ops/s
[info] JsonFramingBenchmark.counting_offer_5        thrpt    5  2794009,199 ± 285732,751  ops/s     → 2849088,721 ± 277487,640  ops/s
```

while the average score seem slightly better in all cases, this is still quite within the last push metrics error intervals. Probably not worth the loss in code legibility; I'll leave it as is.