@cchepelov 2018-10-12
Running with 30 warm-up & 30 iterations yields weird results (as in the initial report):

Master d9a41f5e237a841b7f15e198fb5c22673b6ed1cf at -i 30 -wi 30:
```text
[info] JsonFramingBenchmark.counting_1              thrpt   30  2655804,989 ± 60613,486  ops/s
[info] JsonFramingBenchmark.counting_long_document  thrpt   30      201,483 ±     5,793  ops/s
[info] JsonFramingBenchmark.counting_offer_5        thrpt   30  2285532,218 ± 59965,221  ops/s
```

Updated branch bb456bd7d2c9b9524a35dbfe7c463e7e8967a7d6 at -i 30 -wi 30:
```text
[info] JsonFramingBenchmark.counting_1              thrpt   30  2488869,333 ± 36421,754  ops/s
[info] JsonFramingBenchmark.counting_long_document  thrpt   30      720,803 ±    12,887  ops/s
[info] JsonFramingBenchmark.counting_offer_5        thrpt   30  1458837,532 ± 30467,879  ops/s
```

Now the really weird part is when the next level of JIT kicks in (example on counting_offer_5):
```text
[info] # Warmup Iteration   1: 2510990,608 ops/s
[info] # Warmup Iteration   2: 2526441,393 ops/s
[info] # Warmup Iteration   3: 2641141,125 ops/s
[info] # Warmup Iteration   4: 2659889,369 ops/s
[info] # Warmup Iteration   5: 2737527,462 ops/s
[info] # Warmup Iteration   6: 2638352,978 ops/s
[info] # Warmup Iteration   7: 2551206,206 ops/s
[info] # Warmup Iteration   8: 2791854,785 ops/s
[info] # Warmup Iteration   9: 2771254,012 ops/s
[info] # Warmup Iteration  10: 2774640,490 ops/s
[info] # Warmup Iteration  11: 2714039,178 ops/s
[info] # Warmup Iteration  12: 2623812,107 ops/s
[info] # Warmup Iteration  13: 2631459,200 ops/s
```
so far so good. And then:

```text
[info] # Warmup Iteration  14: 1625617,941 ops/s
[info] # Warmup Iteration  15: 1521581,545 ops/s
[info] # Warmup Iteration  16: 1487617,934 ops/s
[info] # Warmup Iteration  17: 1523624,664 ops/s
[info] # Warmup Iteration  18: 1455953,394 ops/s
[info] # Warmup Iteration  19: 1433586,772 ops/s
[info] # Warmup Iteration  20: 1435661,055 ops/s
[info] # Warmup Iteration  21: 1409643,865 ops/s
[info] # Warmup Iteration  22: 1405667,711 ops/s
[info] # Warmup Iteration  23: 1421820,338 ops/s
[info] # Warmup Iteration  24: 1436581,242 ops/s
[info] # Warmup Iteration  25: 1456347,914 ops/s
[info] # Warmup Iteration  26: 1502848,906 ops/s
[info] # Warmup Iteration  27: 1544719,316 ops/s
[info] # Warmup Iteration  28: 1497289,444 ops/s
[info] # Warmup Iteration  29: 1369107,820 ops/s
[info] # Warmup Iteration  30: 1375654,576 ops/s
[info] Iteration   1: 1401525,462 ops/s
[info] Iteration   2: 1400224,279 ops/s
[info] Iteration   3: 1409271,647 ops/s
[info] Iteration   4: 1456632,194 ops/s
[info] Iteration   5: 1461690,497 ops/s
[info] Iteration   6: 1467002,056 ops/s
[info] Iteration   7: 1436177,597 ops/s
[info] Iteration   8: 1432764,005 ops/s
[info] Iteration   9: 1502544,198 ops/s
```
(this also happens on `counting_1`)

There is no such cliff on master. 
`VM version: JDK 1.8.0_171, OpenJDK 64-Bit Server VM, 25.171-b11`

Clearly something very wrong seems to happen at some stage of JITing… any idea?